requirements/base.txt
Django==6.0.3
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg2-binary==2.9.11
python-decouple==3.8
django-extensions==4.1
django-htmx==1.27.0
